EKIRJASTO-626 Refactor bookDetails component for clarity
